+++ README.md
@@ -1,64 +1,34 @@
 # Twitch Chat Bridge
 
-This light-weight mod links a Twitch channel's chat to your Minecraft chatbox!
+A Fabric mod that allows you to connect your Minecraft chat to a Twitch chat and vise versa.
 
-There are more mods that do this, but none are available for Fabric... so I made it!
+Chat messages from your selected Twitch channel get displayed in your Minecraft chat. Usernames in the MC chat are colored in the users selected color. Also all badges of the user are displayed before their username.
 
-This mod requires [Mod Menu](https://github.com/TerraformersMC/ModMenu) and
-[Fabric API](https://github.com/FabricMC/fabric) to work.
+## Download
 
-**NOTE: Even though the mod is on beta, it's fully functional! It's on beta to symbolise I'm not finished adding features.
-If you encounter any bug or issue or have any suggestion, please add it as an issue
-[here](https://github.com/pblop/twitch-chat/issues/new).**
+Download on [Modrinth](https://modrinth.com/mod/twitch-chat-bridge).
 
-## Usage.
+## Usage
 
-1.  Get a Twitch oauth key here https://twitchapps.com/tmi/ and write it down somewhere.
-2.  Open the mod menu.
-3.  Open this mod's configuration inside that menu.
-4.  Fill in your credentials in the credentials tab (your twitch username, oauth token).
-5.  Go in game and type `/twitch watch CHANNEL` (where CHANNEL should be the name of the Twitch channel you want to join).
-6.  Type `/twitch enable`.
-7.  Now you can talk in the Twitch chat using `:MESSAGE`!
-
-## Commands.
-
-- `/twitch watch CHANNEL` – Changes the watched Twitch channel to `CHANNEL`
-- `/twitch enable` – Starts the Twitch chat integration
-- `/twitch disable` – Stops the Twitch chat integration
-- `/twitch broadcast true` – Relays Twitch chat messages to the Minecraft server as player messages
-- `/twitch broadcast false` – Keeps Twitch chat messages local to the Minecraft client
-
-## Translations.
-
-If you find the mod is not available on a language you know I would really appreciate it if you could create a pull
-request with a translation for that language.
-
-Language files are located in [src/main/resources/assets/twitchchat/lang](src/main/resources/assets/twitchchat/lang)
-and they contain translations for every (translatable) user-facing piece of text in the mod.
-
-If you're interested you'll have to create a file with an appropiate file name for your language. To find your
-language's code, visit this [link](https://minecraft.gamepedia.com/Language#Available_languages).
-
-You should probably use US or UK English as a base for your translation. You can find their respective files
-by clicking on these links: [en_us.json](src/main/resources/assets/twitchchat/lang/en_us.json) and
-[en_gb.json](src/main/resources/assets/twitchchat/lang/en_gb.json).
-
-Don't translate **%d** or **%s** signs, they're used to dynamically insert numbers and text (respectively).
-
-## Contact.
-
-If you have any questions don't hesitate to email, tweet or DM me, you can find my public profiles on my
-[GitHub profile](https://github.com/pblop). I'll answer ASAP.
-
-Just don't personally send me bugs, ideas or feature/help requests, those go in
-[GitHub issues](https://github.com/pblop/twitch-chat/issues).
-
-### Also available on [CurseForge](https://www.curseforge.com/minecraft/mc-mods/twitch-chat) and [ModRinth](https://modrinth.com/mod/twitch-chat).
-
-### [Here](https://github.com/pblop/twitch-chat/wiki/Screenshots) you can see in-game images.
+1. Open the mods config using ModMenu
+2. Go to the credentials tab and fill in your username
+3. Get a token of your twitch account from [twitchtokengenerator.com](https://twitchtokengenerator.com/) and also fill in the token.
+   
+   **NEVER SHARE THIS SITE AFTER LOGGING IN. ALWAYS KEEP ALL YOUR TOKENS AND SECRETS PRIVATE.** They give access to your logged in accout.
+4. Go back ingame and type in the commands:
+   - `/twitch enable` to enable the connection
+   - `/twitch watch CHANNEL` (replace `CHANNEL` with a channel name) to join that channel
+5. You now receive messages of that channels Twitch chat
+   - You can write to that channels chat by typing a `:` at the start of your message (configureable via modmenu)
+   - for example writing the message `:hello guys` sends the message `hello guys` to the joined twitch channel.
 
 ---
 
-I've started school, so this project is going to be updated sparingly, but I'll try to answer every comment and resolve
-your issues.
+## Other
+
+NOTE: Even though the mod is on beta, it's fully functional! It's on beta to symbolise I'm not finished adding features.
+If you encounter any bug or issue or have any suggestion, please add it as an issue
+[here](https://github.com/Kesuaheli/twitch-chat-bridge/issues/new).
+
+This Mod was forked from [pblop/twitch-chat](https://github.com/pblop/twitch-chat) and I added some further additions and bug fixes as well as still improving it.
+Thanks for the Mod base.
