+++ gradle.properties
@@ -12,8 +12,8 @@ org.gradle.parallel=true
 
 # Mod Properties
 	mod_version=0.16.0b
-	maven_group=eu.pabl
-	archives_base_name=twitchchat
+	maven_group=de.kesuaheli
+	archives_base_name=twitchchatbridge
 
 # Dependencies
 	fabric_version=0.111.0+1.21.4
