+++ src/main/java/de/kesuaheli/twitchchatbridge/commands/TwitchBaseCommand.java
@@ -1,4 +1,4 @@
-package eu.pabl.twitchchat.commands;
+package de.kesuaheli.twitchchatbridge.commands;
 
 
 import com.mojang.brigadier.CommandDispatcher;
